+++ ft_printf/src/parse_doubles.c
@@ -6,7 +6,7 @@
 /*   By: skoskine <skoskine@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/09 09:38:46 by skoskine          #+#    #+#             */
-/*   Updated: 2021/05/10 17:51:27 by skoskine         ###   ########.fr       */
+/*   Updated: 2021/05/10 23:42:27 by skoskine         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 #include <stdarg.h>
 #include <stdlib.h>
 
-static int	get_int_part_len(char *value_str)
+static size_t	get_int_part_len(char *value_str)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (value_str[len] != '.' && value_str[len] != '\0')
@@ -51,7 +51,7 @@ char	*parse_double_result(t_data *specs, char *value_str, size_t len)
 	return (result);
 }
 
-int	get_result_length(t_data *specs, double value, char *value_str)
+size_t	get_result_length(t_data *specs, double value, char *value_str)
 {
 	size_t	len;
 
@@ -115,5 +115,5 @@ int	parse_doubles(t_data *specs, va_list *ap, char **result)
 	free(value_str);
 	if (*result == NULL)
 		return (-1);
-	return (len);
+	return ((int)len);
 }
